Move extensions into main files
