+++ lib/src/collection/kt_collection.dart
@@ -1,3 +1,5 @@
+import "dart:math" as math show Random;
+
 import "package:kt_dart/collection.dart";
 
 /// A generic collection of elements. Methods in this interface support only read-only access to the collection;
@@ -21,4 +23,20 @@ abstract class KtCollection<T> implements KtIterable<T> {
   // Bulk Operations
   /// Checks if all elements in the specified collection are contained in this collection.
   bool containsAll(KtCollection<T> elements);
+}
+
+extension KtCollectionExtensions<T> on KtCollection<T> {
+  /// Returns `true` if the collection is not empty.
+  bool isNotEmpty() => size > 0;
+
+  /// Returns a random element from this collection.
+  ///
+  /// @throws NoSuchElementException if this collection is empty.
+  T random([math.Random random]) {
+    final r = random ?? math.Random();
+    return elementAt(r.nextInt(size));
+  }
+
+  /// Returns a [KtMutableList] filled with all elements of this collection.
+  KtMutableList<T> toMutableList() => KtMutableList<T>.from(iter);
 }
